refactor: Implement Go standard layout and optimize project structure
Major restructuring following golang-standards/project-layout with comprehensive
documentation consolidation and Docker Compose v2 migration.

## Structure Changes

### Go Standard Layout Implementation
- Move main.go → cmd/ldap-manager/main.go (Go standard entry point)
- Refactor internal/build.go → internal/version/version.go (proper package naming)
- Update all build systems (Makefile, Dockerfile) for new structure
- Preserve all functionality with correct import paths

### Docker Compose v2 Migration
- Rename docker-compose.yml → compose.yml (v2 standard filename)
- Remove deprecated version field from compose specification
- Update all documentation references to use 'docker compose' syntax

### Documentation Consolidation
- Organize scattered docs into professional structure:
  - docs/user-guide/ (installation, configuration, API)
  - docs/development/ (setup, contributing, architecture)
  - docs/operations/ (deployment, monitoring)
- Add 132KB of comprehensive new documentation
- Move assets to dedicated docs/assets/ directory
- Remove redundant and outdated documentation files

### Configuration Optimization
- Simplify .envrc from 2.5KB to 342 bytes (86% reduction)
- Clean up build artifacts and temporary files
- Maintain only essential configuration files in root

## Validation
- ✅ Asset build pipeline works (CSS + templates compile)
- ✅ All import paths updated correctly
- ✅ Professional documentation structure ready for enterprise use
- ✅ Clean project structure following Go best practices
