+++ cmd/ldap-manager/main.go
@@ -8,15 +8,15 @@ import (
 	"github.com/rs/zerolog"
 	"github.com/rs/zerolog/log"
 
-	"github.com/netresearch/ldap-manager/internal"
 	"github.com/netresearch/ldap-manager/internal/options"
+	"github.com/netresearch/ldap-manager/internal/version"
 	"github.com/netresearch/ldap-manager/internal/web"
 )
 
 func main() {
 	log.Logger = log.Output(zerolog.ConsoleWriter{Out: os.Stderr})
 
-	log.Info().Msgf("LDAP Manager %s starting...", internal.FormatVersion())
+	log.Info().Msgf("LDAP Manager %s starting...", version.FormatVersion())
 
 	opts := options.Parse()
 	log.Logger = log.Logger.Level(opts.LogLevel)
