+++ compose.yml
@@ -1,5 +1,3 @@
-version: "3.8"
-
 services:
   # OpenLDAP server for local development and testing
   openldap:
@@ -92,7 +90,7 @@ services:
       - ldap-network
     restart: unless-stopped
     healthcheck:
-      test: ["CMD", "/ldap-passwd", "--health-check"]
+      test: ["CMD", "/ldap-manager", "--health-check"]
       interval: 30s
       timeout: 3s
       start_period: 5s
