--- coverage-reports/benchmarks.txt
@@ -1,11 +1,0 @@
-?   	github.com/netresearch/ldap-manager	[no test files]
-?   	github.com/netresearch/ldap-manager/internal	[no test files]
-PASS
-ok  	github.com/netresearch/ldap-manager/internal/ldap_cache	0.003s
-PASS
-ok  	github.com/netresearch/ldap-manager/internal/options	0.003s
-PASS
-ok  	github.com/netresearch/ldap-manager/internal/web	0.003s
-?   	github.com/netresearch/ldap-manager/internal/web/static	[no test files]
-PASS
-ok  	github.com/netresearch/ldap-manager/internal/web/templates	0.004s
