--- coverage-reports/summary.txt
@@ -1,3 +1,0 @@
-Coverage Summary:
-Total Coverage: 11.3%
-Threshold: 80%
