--- docs/ANALYSIS.md
@@ -1,150 +1,0 @@
-# Project Analysis Reports
-
-This document consolidates various analysis reports and CI resolution documentation generated during the project development lifecycle.
-
-## Code Quality Refactoring Summary
-
-The LDAP Manager project underwent comprehensive code quality improvements addressing linting violations, complexity issues, and maintainability concerns.
-
-### Issues Fixed
-
-#### 1. Unused Parameters
-
-**Files Modified**: `internal/ldap_cache/test_helpers.go`, `internal/ldap_cache/cache_test.go`, `internal/ldap_cache/manager.go`
-
-**Changes Made**:
-
-- Renamed unused parameters to `_` in mock functions
-- Fixed concurrent test goroutines parameter usage
-- Improved parameter names in Filter functions
-
-#### 2. Function Complexity Reduction
-
-Complex functions were refactored to improve readability and maintainability while staying within cyclomatic complexity limits.
-
-#### 3. Test Coverage Enhancement
-
-All packages now maintain adequate test coverage with comprehensive test suites covering edge cases and error conditions.
-
-## CI/CD Pipeline Resolution
-
-### Executive Summary
-
-All critical CI/CD pipeline failures were systematically investigated and resolved through targeted fixes addressing root causes identified through detailed log analysis.
-
-### Root Cause Analysis Results
-
-#### Primary Issue: Missing templ CLI in CI Environment
-
-**Problem**: `sh: 1: templ: not found` across all workflows
-**Solution**: Added `go install github.com/a-h/templ/cmd/templ@latest` to all jobs requiring asset building
-
-#### Secondary Issue: Formatting Violations
-
-**Problem**: Documentation formatting inconsistencies
-**Solution**: Applied consistent markdown formatting across all documentation files
-
-#### Tertiary Issue: Go Version Inconsistencies
-
-**Problem**: Version mismatches across workflows
-**Solution**: Standardized on Go 1.23+ across all CI environments
-
-### Files Modified
-
-- `.github/workflows/quality.yml` - Added templ CLI installation
-- `.github/workflows/check.yml` - Verified consistency
-- Various documentation files - Applied formatting fixes
-
-### Impact Assessment
-
-- ✅ Template generation failures resolved
-- ✅ Go compilation errors eliminated
-- ✅ CI pipeline stability achieved
-- ✅ Code quality standards maintained
-
-## Test Refactoring Analysis
-
-### Comprehensive Testing Strategy
-
-The project implements a robust testing strategy with multiple layers:
-
-1. **Unit Tests**: Individual component testing with mocking
-2. **Integration Tests**: Cross-component interaction testing
-3. **Benchmark Tests**: Performance regression detection
-4. **Coverage Reporting**: Automated coverage threshold enforcement
-
-### Test Infrastructure Improvements
-
-- **Coverage Threshold**: 80% minimum enforced by CI
-- **Race Detection**: Concurrent safety validation
-- **Benchmark Tracking**: Performance regression alerts
-- **HTML Reports**: Visual coverage analysis
-
-### Quality Metrics
-
-- **Test Coverage**: >80% across all packages
-- **Race Conditions**: Zero detected in concurrent operations
-- **Benchmark Stability**: Consistent performance characteristics
-- **Code Complexity**: All functions under cyclomatic complexity limit of 10
-
-## Development Workflow Optimization
-
-### Build System Enhancements
-
-1. **Asset Processing**: Integrated TailwindCSS and templ compilation
-2. **Hot Reload**: Development server with automatic rebuilds
-3. **Quality Gates**: Pre-commit hooks and CI validation
-4. **Deployment**: Docker containerization with multi-stage builds
-
-### Tool Integration
-
-- **golangci-lint**: Comprehensive static analysis (30+ linters)
-- **govulncheck**: Security vulnerability scanning
-- **gofumpt/goimports**: Code formatting automation
-- **pre-commit**: Git hook integration
-
-### Makefile Automation
-
-The project features a comprehensive Makefile with 40+ targets covering:
-
-- Dependency management and tool installation
-- Building and asset processing
-- Testing with coverage reporting
-- Linting and quality checks
-- Docker operations
-- Development server management
-
-## Architectural Decisions
-
-### Code Organization
-
-- **Package Structure**: Clear separation of concerns (web, ldap_cache, options)
-- **Dependency Injection**: Testable component design
-- **Error Handling**: Consistent error wrapping and logging
-- **Concurrent Safety**: Proper synchronization for shared resources
-
-### Technology Choices
-
-- **Go 1.23+**: Modern language features and performance
-- **Fiber v2**: High-performance web framework
-- **templ**: Type-safe HTML templates
-- **TailwindCSS v4**: Modern CSS framework
-- **PNPM**: Efficient package management
-
-## Future Recommendations
-
-### Continuous Improvement
-
-1. **Performance Monitoring**: Implement application performance monitoring
-2. **Security Scanning**: Regular dependency vulnerability assessment
-3. **Documentation**: Maintain up-to-date API and architecture documentation
-4. **Testing**: Expand integration test coverage for complex workflows
-
-### Technical Debt Management
-
-1. **Code Reviews**: Maintain consistent review processes
-2. **Refactoring**: Regular code quality assessment and improvement
-3. **Dependencies**: Automated dependency updates with testing
-4. **Monitoring**: Production performance and error tracking
-
-This analysis represents the systematic approach taken to maintain code quality, reliability, and maintainability throughout the project lifecycle.
