--- docs/REFACTORING_SUMMARY.md
@@ -1,129 +1,0 @@
-# LDAP Manager Code Quality Refactoring Summary
-
-This document summarizes the code quality and maintainability improvements made to the LDAP Manager project.
-
-## Issues Fixed
-
-### 1. Unused Parameters
-
-**Issue**: Several functions had unused parameters that triggered linter warnings.
-
-**Files Modified**:
-
-- `internal/ldap_cache/test_helpers.go`
-- `internal/ldap_cache/cache_test.go`
-- `internal/ldap_cache/manager.go`
-
-**Changes Made**:
-
-- Renamed unused parameters to `_` in mock functions:
-  - `CheckPasswordForSAMAccountName(_, _ string)`
-  - `WithCredentials(_, _ string)`
-  - `NewMockUser(_ string, samAccountName string, ...)`
-  - `NewMockGroup(_, _ string, members []string)`
-  - `NewMockComputer(_ string, samAccountName string, ...)`
-
-- Fixed concurrent test goroutines to properly use `_` for unused iteration parameters
-- Improved parameter names in Filter functions: `func(u ldap.User)` instead of `func(t ldap.User)`
-
-### 2. Function Complexity Reduction
-
-**Issue**: Two handler functions exceeded complexity threshold (16 > 15):
-
-- `groupModifyHandler`
-- `userModifyHandler`
-
-**Refactoring Strategy**: Extracted common patterns into helper functions to reduce cyclomatic complexity.
-
-#### Groups Handler Refactoring (`internal/web/groups.go`)
-
-**New Helper Functions**:
-
-- `loadGroupData()`: Centralizes group data loading and sorting logic
-- `renderGroupWithError()`: Handles error response rendering
-- `renderGroupWithSuccess()`: Handles success response rendering
-- `performGroupModification()`: Encapsulates LDAP modification operations
-
-**Complexity Reduction**: Reduced from 16 to approximately 8 by:
-
-- Eliminating duplicate data loading code
-- Extracting error handling patterns
-- Separating LDAP operations from HTTP response logic
-
-#### Users Handler Refactoring (`internal/web/users.go`)
-
-**New Helper Functions**:
-
-- `loadUserData()`: Centralizes user data loading and sorting logic
-- `renderUserWithError()`: Handles error response rendering
-- `renderUserWithSuccess()`: Handles success response rendering
-- `performUserModification()`: Encapsulates LDAP modification operations
-
-**Complexity Reduction**: Reduced from 16 to approximately 8 by:
-
-- Eliminating duplicate data loading code
-- Extracting error handling patterns
-- Separating LDAP operations from HTTP response logic
-
-### 3. Test Code Quality
-
-**File**: `internal/web/handlers_test.go`
-
-- Properly using `_` for unused return values from `setupTestApp()` (already correct)
-
-**File**: `internal/ldap_cache/cache_test.go`
-
-- Fixed concurrent test goroutines to use `_` for unused iteration parameters
-- Simplified Filter test logic to focus on concurrent access rather than data matching
-
-### 4. Code Structure Improvements
-
-**File**: `internal/ldap_cache/manager.go`
-
-- Improved parameter naming in Filter functions for better readability
-- Added missing blank line before return statement in `FindComputers()`
-
-## Quality Metrics Improvement
-
-### Before Refactoring
-
-- `groupModifyHandler`: Complexity 16
-- `userModifyHandler`: Complexity 16
-- Multiple unused parameter warnings
-- Code duplication in error handling and data loading
-
-### After Refactoring
-
-- `groupModifyHandler`: Complexity ~8 (50% reduction)
-- `userModifyHandler`: Complexity ~8 (50% reduction)
-- Zero unused parameter warnings
-- DRY principle applied with helper functions
-- Improved separation of concerns
-- Better testability through smaller, focused functions
-
-## Benefits Achieved
-
-1. **Maintainability**: Smaller functions are easier to understand, test, and modify
-2. **Code Reuse**: Common patterns extracted to helper functions eliminate duplication
-3. **Readability**: Clearer function names and separation of concerns
-4. **Error Handling**: Consistent error response patterns
-5. **Testing**: Helper functions can be unit tested independently
-6. **Linting**: Eliminated all unused parameter warnings
-
-## Technical Debt Reduction
-
-- **Duplication**: Eliminated repeated data loading and response rendering code
-- **Complexity**: Reduced cyclomatic complexity to maintainable levels
-- **Naming**: Improved parameter and function naming conventions
-- **Structure**: Better separation between HTTP handling and business logic
-
-## Preserved Functionality
-
-All refactoring changes maintain 100% backward compatibility:
-
-- No changes to public APIs or interfaces
-- All existing functionality preserved
-- Same error handling behavior
-- Identical user experience
-
-The refactoring focused purely on internal code structure improvements without altering external behavior.
