--- docs/architecture.md
@@ -1,4 +1,0 @@
-# Architecture
-
-The following architecture will be used:
-![Architecture](./architecture.png)
