+++ docs/assets/architecture.png
