+++ docs/assets/ldap_manager_form.png
