+++ docs/assets/ldap_manager_form_errors.png
